Make Jahia compile on Java 1.5 (almost one year this version is dead in oracle eyes (end of life Nov 2009))
git-svn-id: svn://localhost/jahia/trunk@7821 69c369ed-ed09-0410-a4a1-e19a46c23626

+++ modules/rating/src/main/java/org/jahia/modules/rating/mahout/Recommender.java
@@ -131,7 +131,7 @@ public class Recommender implements Processor, InitializingBean, CamelContextAwa
                         resultSet.next();
                         long value = resultSet.getLong(1);
                         paths.put(path, value);
-                        invertedPaths.put((long) value, path);
+                        invertedPaths.put(value, path);
                         resultSet.close();
                         statement.close();
                     }
